% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health3_Statistics.R
\name{afl}
\alias{afl}
\title{Average future lifetime}
\usage{
afl(init_age, init_state, trans_probs = NULL, simulated_path = NULL)
}
\arguments{
\item{init_age}{integer between 65 and 110 denoting initial age of individual. This needs to be same
initial age used in generation of \code{trans_probs} or \code{simulated_path}}

\item{init_state}{0 for healthy, 1 for disabled}

\item{trans_probs}{a list of transition probability matrices, preferably generated from
\code{\link[tshm]{get_trans_probs}}.}

\item{simulated_path}{a matrix of lifetime paths simulations, generated by \code{\link[tshm]{simulate_path}}.}
}
\value{
numeric output for average and standard deviation of future lifetime
}
\description{
Calculates the average future life time and its standard deviation
given initial state and age of an individual by simulating life time paths.
Function needs to be provided either a list of transition probability matrices,
or a simulated lifetime path from \code{\link[tshm]{simulate_path}}. If both are provided,
then the function will use the simulated path provided.
NOTE: USE \code{\link[tshm]{aflF}} for frailty model.
}
