% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pricing_Cashflow.R
\name{cf_account_based_pension}
\alias{cf_account_based_pension}
\title{Account Based Pension - Cashflow Simulator}
\usage{
cf_account_based_pension(policy, state, data)
}
\arguments{
\item{policy}{Policy object containing necessary parameters (see create_policy_AP)}

\item{state}{State vector containing state values for entire duration}

\item{data}{Data frame containing all variables generated using other modules}
}
\value{
Vector of cashflows for at each time point
}
\description{
Simulate cash flows for Account Based Pension
}
