% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pricing_Cashflow.R
\name{cf_care_annuity}
\alias{cf_care_annuity}
\title{Care Annuity - Cashflow Simulator}
\usage{
cf_care_annuity(policy, state, data)
}
\arguments{
\item{policy}{Policy object containing necessary parameters (see create_policy_CA)}

\item{state}{State vector containing state values for entire duration}

\item{data}{Data frame containing all variables generated using other modules}
}
\value{
Vector of cashflows for at each time point
}
\description{
Care Annuity - Cashflow Simulator
}
