% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pricing_Cashflow.R
\name{cf_reverse_mortgage}
\alias{cf_reverse_mortgage}
\title{Reverse Mortgage - Cashflow Simulator}
\usage{
cf_reverse_mortgage(policy, state, data)
}
\arguments{
\item{policy}{Policy object containing necessary parameters (see create_policy_RM)}

\item{state}{State vector containing state values for entire duration}

\item{data}{Data frame containing all variables generated using other modules}
}
\value{
Vector of cashflows for at each time point
}
\description{
Reverse Mortgage - Cashflow Simulator
}
