% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health_Data.R
\docType{data}
\name{china_CLHLS}
\alias{china_CLHLS}
\title{Parameters of 3-state Cox Regression Model estimated from CLHLS}
\format{
A data frame with 15 rows and 6 variables:
\describe{
\item{Model}{static, trend or frailty}
\item{Coef}{the covariate that the coefficient is corresponding to}
\item{trans1}{parameter estimate for H to M transition}
\item{trans2}{parameter estimate for H to D transition}
\item{trans3}{parameter estimate for H to MD transition}
\item{trans4}{parameter estimate for H to Dead transition}
}
}
\source{
\url{https://cepar.edu.au/publications/working-papers/functional-disability-systematic-trends-and-uncertainty-comparison-between-china-and-us}
}
\usage{
china_CLHLS
}
\description{
A dataset containing the parameters of the 3-state static, trend and frailty cox regression
models estimated with the 1998-2014 CLHLS data.
}
\keyword{datasets}
