% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_Utility.R
\name{cohort2period}
\alias{cohort2period}
\title{Convert from Cohort to Period Rates}
\usage{
cohort2period(cohort_rates)
}
\arguments{
\item{cohort_rates}{matrix or 3D array of mortality rates with age (on the rows) and cohort
(on the columns) and simulation number (3rd dimension)}
}
\value{
associated period mortality rates as a 3D array if \code{cohort_rates} is an array,
or as a matrix otherwise
}
\description{
Converts from cohort to period mortality rates
}
