% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_Utility.R
\name{combine_hist_sim}
\alias{combine_hist_sim}
\title{Combine Historical and Simulated Rates}
\usage{
combine_hist_sim(rates_hist, rates_sim)
}
\arguments{
\item{rates_hist}{Matrix of historical mortality rates with age (on the rows) and
calendar year or cohort (on the columns)}

\item{rates_sim}{3D array of simulated mortality rates with age (on the rows) and
calendar year or cohort (on the columns) and simulation number (3rd dimension)}
}
\value{
3D array of combined historical and simulated rates
}
\description{
Combine Historical and Simulated Rates
}
