% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health_Wrapper.R
\name{create_life_table}
\alias{create_life_table}
\title{Create life table}
\usage{
create_life_table(trans_probs, init_age, init_state = 0, cohort = 1e+05)
}
\arguments{
\item{trans_probs}{list of transition probability matrices; typically generated by \code{get_trans_probs}}

\item{init_age}{integer denoting current age}

\item{init_state}{integer value
for 3-state model: integer value of 0 or 1, where 0 for healthy state, 1 for disabled state
for 5-state model: 0 for H state, 1 for M state, 2 for D state, 3 for MD state}

\item{cohort}{initial cohort size for lifetable}
}
\value{
dataframe of lifetable generated
}
\description{
Creates life table from a list of transition probability matrices at some initial age
and specified cohort size.
}
\examples{
example

}
