% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pricing_Policy.R
\name{create_policy_AP}
\alias{create_policy_AP}
\title{Account Based Pension - Policy Object Generator}
\usage{
create_policy_AP(balance, expenses)
}
\arguments{
\item{balance}{Initial balance of Pension account}

\item{expenses}{Initial yearly expenses for policyholder}
}
\value{
Policy object
}
\description{
Generates policy object for Account Based Pension policies
}
\examples{
ap <- create_policy_AP(1000000, 50000)
}
