% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pricing_Policy.R
\name{create_policy_CA}
\alias{create_policy_CA}
\title{Care Annuity - Policy Object Generator}
\usage{
create_policy_CA(benefit, increase, min, loading)
}
\arguments{
\item{benefit}{Annual benefit for each policy (vector)}

\item{increase}{Annual rate of increase for each policy (vector)}

\item{min}{Minimum guaranteed period for each policy (vector)}

\item{loading}{(Optional) Loading factor for contract}
}
\value{
Policy object
}
\description{
Generates policy object for Care Annuity policies
}
\examples{
ca <- create_policy_CA(c(60000, 1200), c(0, 0.04), c(5, 0), c(0.04, 0.01))
}
