% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pricing_Policy.R
\name{create_policy_LA}
\alias{create_policy_LA}
\title{Life Annuity - Policy Object Generator}
\usage{
create_policy_LA(benefit, defer = 0, increase = 0, loading)
}
\arguments{
\item{benefit}{Annual benefit for policy}

\item{defer}{(Optional) Deferment period of policy}

\item{increase}{(Optional) Annual rate of increase for policy}

\item{loading}{(Optional) Loading factor for contract}
}
\value{
Policy object
}
\description{
Generates policy object for Life Annuity  policies
}
\examples{
la <- create_policy_LA(60000, 5, 0, 0.01)
}
