% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pricing_Policy.R
\name{create_policy_PA}
\alias{create_policy_PA}
\title{Pooled Annuity - Policy Object Generator}
\usage{
create_policy_PA(benefit, size, interest, loading)
}
\arguments{
\item{benefit}{Initial annual benefit for policy}

\item{size}{Size of pool}

\item{interest}{Assumed investment earnings rate (flat)}

\item{loading}{(Optional) Loading factor for contract}
}
\value{
Policy object
}
\description{
Generates policy object for Pooled Annuity policies
}
\examples{
pa <- create_policy_PA(60000, 10000, 0.05, 0.01)
}
