% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pricing_Policy.R
\name{create_policy_RM}
\alias{create_policy_RM}
\title{Reverse Mortgage - Policy Object Generator}
\usage{
create_policy_RM(value, LVR, trans_cost, margin)
}
\arguments{
\item{value}{Initial value of property}

\item{LVR}{Loan to Value ratio for PH}

\item{trans_cost}{Transaction cost associated with sale}

\item{margin}{Lending margin}
}
\value{
Policy object
}
\description{
Generates policy object for Reverse Mortgage policies
}
\examples{
rm <- create_policy_RM(600000, 0.64, 0.01, 0.04)
}
