% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pricing_Policy.R
\name{create_policy_VA}
\alias{create_policy_VA}
\title{Variable Annuity - Policy Object Generator}
\usage{
create_policy_VA(value, length, prop, g_fee)
}
\arguments{
\item{value}{Initial Account value (single upfront premium)}

\item{length}{Length of contract (in years)}

\item{prop}{Proportion of the premium that can be withdrawn annually}

\item{g_fee}{Continuous guarantee fees for GMWB}
}
\value{
Policy object
}
\description{
Generates policy object for Variable Annuity policies
}
\examples{
va <- create_policy_VA(1000000, 30, 0.1, 0.01)
}
