% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/economic_AFNS.R
\name{esg_afns_simulation}
\alias{esg_afns_simulation}
\title{esg_afns_simulation}
\usage{
esg_afns_simulation(
  num_years = 5,
  num_paths = 10,
  frequency = "month",
  perc_change = FALSE,
  type = "independent",
  model = "interest_rate"
)
}
\arguments{
\item{num_years}{Number of years to forecast, counting from 2021-06-01. Default is 5 years.}

\item{num_paths}{Number of simulation paths. Default is 10 paths.}

\item{frequency}{One of "year", "quarter", and "month" (default).}

\item{perc_change}{If the outputs are expressed in terms of period-by-period percentage
change.Default is FALSE. The reference level, i.e., the original values in the first output period, will be appended above the percentage changes for each variable and each trajectory.}

\item{type}{Either "independent" (default) or "correlated". Independent-factor
model assumes independence between the latent factors for interest rates.}

\item{model}{Either "interest_rate" (default) or "interest_house_stock".
The latter model is based on an augmented Arbitrage-Free Nelson-Siegel model,
where both home value index and stock price are assumed to be dependent on interest rates.}
}
\value{
If model is \code{interest_rate}, the function returns a list containing
40 data frames for the simulated trajectories for maturities from 1 quarter up to 10 years.
If model is \code{interest_house_stock},the function returns a list containing 42 data frames
for the simulated trajectories for maturities from 1 quarter up to 10 years,
as well as NSW house value indexes and S&P/ASX200 closing prices.
}
\description{
Returns the simulated paths of the zero-coupon interest rate term structure.
The model is based on an Arbitrage-Free Nelson-Siegel (AFNS) model.
This function uses the package \code{MASS} for random number generation and \code{expm}
for matrix exponential.
}
\examples{
sim = esg_afns_simulations(num_years = 10, num_paths = 100, 
frequency = "year", type = "correlated", model = "interest_rate"). To obtain trajectories of 
Australia 3-month zero-coupon yields, type sim$maturity_1qtrs. To obtain trajectories of 
S&P/ASX200 closing prices, type sim$stock_price. 
}
