% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_Statistics.R
\name{exp_cfl}
\alias{exp_cfl}
\title{Summarise Curtate Future Lifetime Statistics}
\usage{
exp_cfl(qx, ages, init_age = NULL, years = NULL)
}
\arguments{
\item{qx}{vector, matrix or 3D array of 1-year death probabilities with age
(on the rows) and calendar year or cohort (on the columns) and simulation
number (3rd dimension)}

\item{ages}{vector of ages for \code{qx}}

\item{init_age}{initial age for which the curtate future lifetime is to be calculated at. If not provided,
the summary statistics will be calculated for the smallest age supplied in \code{ages}}

\item{years}{optional vector of years for \code{qx}. If not supplied, then the column names
of \code{qx} will be preserved}
}
\value{
expected curtate future lifetime as a matrix if \code{qx} is a 3D array
with simulation number (on the rows) and calendar year or cohort (on the columns).
Returns a vector otherwise
}
\description{
Produces expected curtate future lifetime for a 3D array of 1-year
death probabilities.
}
