% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health5_Transition.R
\name{get_full_trans_prob_matrix}
\alias{get_full_trans_prob_matrix}
\title{the function to get a full list of transition probability matrices from the initial age to age 110}
\usage{
get_full_trans_prob_matrix(params, init_age, gender, i, model)
}
\arguments{
\item{params}{the input parameters from the estimation results, read from an excel file}

\item{init_age}{the initial age of the transition probability matrices}

\item{gender}{gender 1 if female, 0 if male}

\item{i}{the wave index}

\item{model}{1 for no-frailty model, 2 for no-frailty model with a trend, 3 for frailty model}
}
\value{
a list of 5 times 5 transition probability matrices, from the initial age to age 110
}
\description{
the function to get a full list of transition probability matrices from the initial age to age 110
}
\examples{
trans_prob_matrix_age65to110=get_full_trans_prob_matrix(params, init_age=65, gender, i, model=3)
}
