% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pricing_Valuation.R
\name{get_path_prices}
\alias{get_path_prices}
\title{Valuation of Cashflows}
\usage{
get_path_prices(cashflows)
}
\arguments{
\item{cashflows}{Matrix of simulated cashflow paths}
}
\value{
Policy price
}
\description{
Calculates value of a policy using a provided set of cashflows
}
