% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pricing_Simulator.R
\name{get_policy_scenario}
\alias{get_policy_scenario}
\title{Scenario Generator}
\usage{
get_policy_scenario(policy, age, sex, seed, n, period, econ_var)
}
\arguments{
\item{policy}{Policy object containing necessary parameters (see create_policy_ )}

\item{age}{Initial age of policyholder in years}

\item{sex}{Sex of policyholder}

\item{seed}{Seed for random generator}

\item{n}{Number of paths to simulate (Monte-Carlo method)}

\item{period}{Number of periods to simulate}

\item{econ_var}{Simulated economic variables via Economic Scenario Generator}
}
\value{
Matrix of cash flow vectors for each simulated path

Data frame containing all variables generated using other modules
}
\description{
Generates and encapsulates scenario data for a given policy
}
