% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pricing_Valuation.R
\name{get_price_stats}
\alias{get_price_stats}
\title{Calculates statistics for Cashflows}
\usage{
get_price_stats(prices)
}
\arguments{
\item{prices}{Matrix of simulated cashflow paths}
}
\value{
List of statistics
}
\description{
Calculates general statistics for a provided set of cashflows
}
