% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health_Wrapper.R
\name{get_trans_probs}
\alias{get_trans_probs}
\title{Calculate transition probability matrices}
\usage{
get_trans_probs(
  n_states,
  model_type,
  param_file,
  init_age,
  female,
  year = 2012,
  wave_index = 8,
  latent = 0
)
}
\arguments{
\item{n_states}{take values 3 or 5, use 3 for 3-state model, and 5 for 5-state model}

\item{model_type}{string that selects model type; S for Static, T for Trend and F for Frailty}

\item{param_file}{string for file path of parameter file OR a tibble/dataframe of parameters}

\item{init_age}{integer denoting age of policy holder}

\item{female}{takes values 1 or 0, where 1 indicates policyholder is female}

\item{year}{integer denoting current year}

\item{wave_index}{the wave index = (interview year - 1998)/2 + 1}

\item{latent}{initial value of latent factor, normally take the value 0}
}
\value{
list of transition probability matrices
}
\description{
Creates a list of transition probability matrices starting from initial age to
max age of 110 for 3 state model or 5 state model
}
\examples{
trans_probs=get_trans_probs(n_states=5, model_type='F', 
param_file=US_HRS_5, init_age=65, female=0, year = 2012, wave_index = 8, 
latent = 0)

}
