% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health3_Statistics.R
\name{health3_afl}
\alias{health3_afl}
\title{Average future lifetime}
\usage{
health3_afl(
  model_type,
  init_age,
  init_state,
  trans_probs = NULL,
  simulated_path = NULL,
  female = NULL,
  year = NULL,
  param_file = NULL,
  n = 1000
)
}
\arguments{
\item{model_type}{S for static model, T for trend model, F for frailty model}

\item{init_age}{integer between 65 and 110 denoting initial age of individual. This needs to be same
initial age used in generation of \code{trans_probs} or \code{simulated_path}}

\item{init_state}{0 for healthy, 1 for disabled}

\item{trans_probs}{a list of transition probability matrices, preferably generated from
\code{get_trans_probs}, only needed for static and trend models.}

\item{simulated_path}{a matrix of lifetime paths simulations, generated by \code{simulate_health_state_paths}, only needed for static and trend models.}

\item{female}{0 for male, 1 for female, compulsory variable for frailty model}

\item{year}{integer indicating current year, compulsory variable for frailty model}

\item{param_file}{string for file path of parameter file OR a tibble/dataframe of parameters, compulsory variable for frailty model}

\item{n}{integer denoting number of unique latent factor simulations}
}
\value{
numeric output for average and standard deviation of future lifetime
}
\description{
Calculates the average future life time and its standard deviation
given initial state and age of an individual by simulating life time paths.
Function needs to be provided either a list of transition probability matrices,
or a simulated lifetime path from \code{simulate_health_state_paths}. If both are provided,
then the function will use the simulated path provided.
}
\examples{
example
}
