% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health3_Statistics.R
\name{health3_aflF}
\alias{health3_aflF}
\title{Average future lifetime (Frailty Model)}
\usage{
health3_aflF(init_age, init_state, female, year, param_file, n = 1000)
}
\arguments{
\item{init_age}{integer between 65 and 110 indicating initial age of individual}

\item{init_state}{0 for healthy, 1 for disabled}

\item{female}{0 for male, 1 for female}

\item{year}{integer indicating current year}

\item{param_file}{string for file path of parameter file OR a tibble/dataframe of parameters}

\item{n}{integer denoting number of unique latent factor simulations}
}
\value{
numeric output for expected future lifetime and standard deviation of future lifetime
}
\description{
Performs the same function as the \code{health3_afl}, but simulates a number of
unique latent factor paths to make probabilities stochastic.
NOTE: USE \code{health3_afl} for static and trend models.
}
\examples{
example
}
