% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health3_LifeTable.R
\name{health3_create_life_table}
\alias{health3_create_life_table}
\title{Create life table}
\usage{
health3_create_life_table(trans_probs, init_age, init_state, cohort)
}
\arguments{
\item{trans_probs}{list of transition probability matrices; typically generated by \code{health3_get_trans_probs}}

\item{init_age}{integer denoting current age}

\item{init_state}{integer value of 0 or 1, where 0 indicates healthy and 1 indicates disabled}

\item{cohort}{initial cohort size for lifetable}
}
\value{
dataframe of lifetable generated
}
\description{
Creates life table from a list of transition probability matrices at some initial age
and specified cohort size.
}
\examples{
example

}
