% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health3_Statistics.R
\name{health3_dflF}
\alias{health3_dflF}
\title{Disabled Future Lifetime (frailty model)}
\usage{
health3_dflF(init_age, init_state, female, year, param_file, n = 1000)
}
\arguments{
\item{init_age}{integer between 65 and 110 denoting initial age of individual}

\item{init_state}{0 for healthy and 1 for disabled}

\item{female}{0 for non female, 1 for female}

\item{year}{integer for current year}

\item{param_file}{string for file path of parameter file OR a tibble/dataframe of parameters}

\item{n}{integer representing number of unique latent factors to simulate}
}
\value{
numeric output for average time spent and standard deviation of time in disabled state
}
\description{
Performs the same function as \code{health3_dfl}, but it requires inputs
to simulate latent factor in the frailty model. This allows it to simulate the
the randomness from the frailty factor into the calculation as well.
NOTE: THIS FUNCTION SHOULD ONLY BE USED WITH THE FRAILTY MODEL. USE \code{health3_dfl}
FOR STATIC AND TREND MODEL.
}
\examples{
example
}
