% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health3_Statistics.R
\name{health3_hflF}
\alias{health3_hflF}
\title{Heatlhy Future Lifetime (Frailty Model)}
\usage{
health3_hflF(init_age, init_state, female, year, param_file, n = 1000)
}
\arguments{
\item{init_age}{integer between 65 and 110 denoting initial age of individual}

\item{init_state}{0 for healthy, 1 for disabled}

\item{female}{0 for male, 1 for female}

\item{year}{integer denoting current year}

\item{param_file}{string for file path of parameter file OR a tibble/dataframe of parameters}

\item{n}{integer denoting number of simulations}
}
\value{
Numeric output denoting the average time spent in healthy state and its
standard deviation.
}
\description{
Performs the same function as \code{health3_hfl}, but simulates unique latent factor
paths in the process.
NOTE: USE \code{health3_hfl} for static and trend models.
}
\examples{
example
}
