% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health3_Plot.R
\name{health3_prob_plots}
\alias{health3_prob_plots}
\title{Create Survival Probability Plots}
\usage{
health3_prob_plots(init_state, init_age, trans_probs)
}
\arguments{
\item{init_state}{0 for healthy, 1 for disabled.}

\item{init_age}{integer between 65 and 110 (inclusive) denoting initial age of individual. This has
to be the same as the initial age used for transition probability matrices simulation.}

\item{trans_probs}{list of transition probability matrices, this could be generated by \code{get_trans_probs}.}
}
\value{
line plot with 3 lines, each denoting survival probability, survival probability to
disabled, and survival probability to healthy.
}
\description{
Create a plot of survival probabilities to alive, health, and disabled states.
}
\examples{
example
}
