% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health3_Plot.R
\name{health3_surv_prob}
\alias{health3_surv_prob}
\title{Calculate Survival Probabilities}
\usage{
health3_surv_prob(
  init_state,
  init_age,
  target_age,
  trans_probs,
  end_state = NULL
)
}
\arguments{
\item{init_state}{0 (for healthy) or 1 (for disabled)}

\item{init_age}{integer between 65 and 110 denoting age of individual. This has to be the same initial
age used in the calculation of transition probability matrices.}

\item{target_age}{integer between 65 and 110 denoting the target age to survive to.}

\item{trans_probs}{list of transition probability matrices; this could be generated by \code{health3_get_trans_probs}.}

\item{end_state}{default NULL to return probability of surviving. Set to 0 for probability ending is healthy, and 1 for disabled.}
}
\value{
numeric denoting probability of surviving to end age in end state
}
\description{
Calculates the probability of surviving to a certain age from an initial state and age
}
\examples{
example
}
