% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health3_Statistics.R
\name{health3_survival_stats}
\alias{health3_survival_stats}
\title{Survival Statistics}
\usage{
health3_survival_stats(
  model_type,
  init_age,
  init_state,
  trans_probs = NULL,
  simulated_path = NULL,
  female = NULL,
  year = NULL,
  param_file = NULL,
  n = 1000
)
}
\arguments{
\item{model_type}{S for static model, T for trend model, F for frailty model}

\item{init_age}{integer between 65 and 110 denoting initial age of individual}

\item{init_state}{0 for healthy, 1 for disabled}

\item{trans_probs}{list of transition probability matrices, generated from \code{health3_get_trans_probs}.}

\item{simulated_path}{matrix containing lifetime path simulations from \code{simulate_health_state_paths} function.}

\item{female}{0 for male, 1 for female, compulsory variable for frailty model}

\item{year}{integer indicating current year, compulsory variable for frailty model}

\item{param_file}{string for file path of parameter file OR a tibble/dataframe of parameters, compulsory variable for frailty model}

\item{n}{integer denoting number of unique latent factor simulations}
}
\value{
dataframe output containing mean and standard deviation of different statistics
}
\description{
Produces statistics including: total expected lifetime, healthy lifetime,
disabled lifetime, onset of disability (if initial state is healthy). It uses
either transition probability matrices or a simulated paths.
NOTE: USE \code{health3_survival_statsF} FOR FRAILTY MODEL.
}
\examples{
example
}
