% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health3_Statistics.R
\name{health3_survival_statsF}
\alias{health3_survival_statsF}
\title{Survival Statistics (Frailty Model)}
\usage{
health3_survival_statsF(
  init_age,
  init_state,
  female,
  year,
  param_file,
  n = 1000
)
}
\arguments{
\item{init_age}{integer between 65 and 110 denoting initial age of individual}

\item{init_state}{0 for healthy, 1 for disabled}

\item{female}{0 for male, 1 for female}

\item{year}{integer denoting current year}

\item{param_file}{string for file path of parameter file OR a tibble/dataframe of parameters}

\item{n}{integer denoting number of latent factor simulations}
}
\value{
dataframe output containing mean and standard deviation of different statistics
}
\description{
Produces statistics including: total expected lifetime, healthy lifetime,
disabled lifetime, onset of disability (if initial state is healthy). It uses
either transition probability matrices or a simulated paths.
}
\examples{
example
}
