% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health3_Statistics.R
\name{health3_time_to_disabled}
\alias{health3_time_to_disabled}
\title{Time until onset of disability (conditional on being disabled)}
\usage{
health3_time_to_disabled(
  model_type,
  init_age,
  init_state,
  trans_probs = NULL,
  simulated_path = NULL,
  female = NULL,
  year = NULL,
  param_file = NULL,
  n = 1000
)
}
\arguments{
\item{model_type}{S for static model, T for trend model, F for frailty model}

\item{init_age}{integer between 65 and 110 denoting initial age of individual}

\item{init_state}{0 for healthy, 1 for disabled}

\item{trans_probs}{a list of transition probability matrices, preferably generated from
\code{health3_get_trans_probs}, only needed for static and trend models.}

\item{simulated_path}{a matrix of lifetime paths simulations, generated by \code{simulate_health_state_paths}, only needed for static and trend models.}

\item{female}{0 for male, 1 for female, compulsory variable for frailty model}

\item{year}{integer indicating current year, compulsory variable for frailty model}

\item{param_file}{string for file path of parameter file OR a tibble/dataframe of parameters, compulsory variable for frailty model}

\item{n}{integer denoting number of unique latent factor simulations}
}
\value{
numeric denoting the average time until individual becomes disabled and its
standard deviation.
}
\description{
Uses simulation to produce an average time until a healthy individual becomes disabled
during their life time. Note that this is conditional on the individual becoming disabled,
so it doesn't factor in life simulations where the individual is always healthy. Function
requires either list of transition probability matrices, or simulated lifetime paths.
If both are given, then the simulated lifetime paths is used.
}
\examples{
example
}
