% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health3_Statistics.R
\name{health3_time_to_disabledF}
\alias{health3_time_to_disabledF}
\title{Time until onset of disability (Frailty model)}
\usage{
health3_time_to_disabledF(init_age, female, year, param_file, n = 1000)
}
\arguments{
\item{init_age}{integer between 65 and 110}

\item{female}{0 for male, 1 for female}

\item{year}{integer denoting current year}

\item{param_file}{string for file path of parameter file OR a tibble/dataframe of parameters}

\item{n}{integer denoting number of simulations to make}
}
\value{
numeric denoting the average time until individual becomes disabled, and its
standard deviation.
}
\description{
Function serves the same purpose as \code{health3_time_to_disabled}, but it is for
the frailty model.
}
\examples{
example
}
