% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health5_Simulator.R
\name{health5_get_life_table}
\alias{health5_get_life_table}
\title{the function to get n_sim number of lifetables}
\usage{
health5_get_life_table(
  model,
  params,
  init_age,
  gender,
  i,
  latent,
  initial_state,
  n_sim = 100
)
}
\arguments{
\item{model}{1 for no-frailty model, 2 for no-frailty model with a trend, 3 for frailty model}

\item{params}{matrix of estimated parameters to construct the five state model. The rows are beta, gamma_age,
gamma_f, phi (if trend or frailty model), alpha (if frailty model). The columns are 1-12 transition types.}

\item{init_age}{the initial age of the life table}

\item{gender}{gender 1 if female, 0 if male}

\item{i}{the wave index = (interview year - 1998)/2 + 1}

\item{latent}{initial value of latent factor, normally take the value 0}

\item{initial_state}{0 for H state, 1 for M state, 2 for D state, 3 for MD state}

\item{n_sim}{the number of simulations}
}
\value{
a list of n_sum number of life table matrices
for each matrix, the row represents the age from the input initial age to 110, and the columns are states H M D MD Dead
for model 3 the frailty model, it simulates the latent factor to get n_sim number of lifetables, so we can get a distribution of the elements in the lifetable
for model 1 and 2, n_sim is suggest to set to be 1 to get one lifetable, otherwise it will produce the same lifetable n_sim times
}
\description{
the function to get n_sim number of lifetables
}
\examples{
simulated_lifetable=health5_get_life_table(model='F', params=params_5_frailty,init_age=65,gender=0,i=8,latent=0,initial_state=0,n_sim=100)
}
