% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health5_Transition.R
\name{health5_get_list_trans_prob_matrix}
\alias{health5_get_list_trans_prob_matrix}
\title{the function to get a full list of transition probability matrices from the initial age to age 110}
\usage{
health5_get_list_trans_prob_matrix(model, params, init_age, gender, i)
}
\arguments{
\item{model}{S for static model, T for trend model, F for frailty model}

\item{params}{matrix of estimated parameters to construct the five state model. The rows are beta, gamma_age,
gamma_f, phi (if trend or frailty model), alpha (if frailty model). The columns are 1-12 transition types.}

\item{init_age}{the initial age of the transition probability matrices}

\item{gender}{gender 1 if female, 0 if male}

\item{i}{the wave index}
}
\value{
a list of 5 times 5 transition probability matrices, from the initial age to age 110
}
\description{
the function to get a full list of transition probability matrices from the initial age to age 110
}
\examples{
trans_prob_matrix_age65to110=health5_get_list_trans_prob_matrix(model='F', params=params_5_frailty, init_age=65, gender=0, i=8)
}
