% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health5_Transition.R
\name{health5_get_trans_probs}
\alias{health5_get_trans_probs}
\title{the function to get a full list of transition probability matrices from the initial age to age 110}
\usage{
health5_get_trans_probs(
  model_type,
  param_file,
  init_age,
  female,
  wave_index,
  latent = 0
)
}
\arguments{
\item{param_file}{matrix of estimated parameters to construct the five state model. The rows are beta, gamma_age,
gamma_f, phi (if trend or frailty model), alpha (if frailty model). The columns are 1-12 transition types.}

\item{init_age}{the initial age of the transition probability matrices}

\item{female}{female 1 if female, 0 if male}

\item{wave_index}{the wave index}

\item{latent}{initial value of latent factor, normally take the value 0}

\item{model}{S for static model, T for trend model, F for frailty model}
}
\value{
a list of 5 times 5 transition probability matrices, from the initial age to age 110
}
\description{
the function to get a full list of transition probability matrices from the initial age to age 110
}
\examples{
trans_prob_matrix_age65to110=health5_get_trans_probs(model_type='F', param_file=params_5_frailty, 
init_age=65, female=0, wave_index=8)
}
