% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health5_Transition.R
\name{health5_get_trans_probs_at_age}
\alias{health5_get_trans_probs_at_age}
\title{function to get a matrix of transition probabilities at a certain age}
\usage{
health5_get_trans_probs_at_age(
  model_type,
  param_file,
  age,
  female,
  wave_index,
  latent
)
}
\arguments{
\item{model_type}{S for static model, T for trend model, F for frailty model}

\item{param_file}{matrix of estimated parameters to construct the five state model. The rows are beta, gamma_age,
gamma_f, phi (if trend or frailty model), alpha (if frailty model). The columns are 1-12 transition types.}

\item{age}{age of the individual}

\item{female}{female 1 if female, 0 if male}

\item{wave_index}{the wave index}

\item{latent}{initial value of latent factor, normally take the value 0}
}
\value{
5 times 5 matrix of transitions probabilities, the states are H M D MD Dead on the rows and columns
}
\description{
function to get a matrix of transition probabilities at a certain age
}
\examples{
transition_probabilities=health5_get_trans_probs_at_age(model_type='F', param_file=params_5_frailty, 
age=65, female=0, wave_index=8, latent=0)
}
