% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health5_Transition.R
\name{health5_get_trans_rates}
\alias{health5_get_trans_rates}
\title{function to calculate transition rates at a certain age}
\usage{
health5_get_trans_rates(
  model_type,
  param_file,
  age,
  female,
  wave_index,
  latent
)
}
\arguments{
\item{model_type}{S for static model, T for trend model, F for frailty model}

\item{param_file}{matrix of estimated parameters to construct the five state model. Generally,
use US_HRS_5 for 5 state model.}

\item{age}{age of the individual}

\item{female}{female 1 if female, 0 if male}

\item{wave_index}{the wave index = (interview year - 1998)/2 + 1}

\item{latent}{initial value of latent factor, normally take the value 0}
}
\value{
12 times 1 vector of transition rates for the 12 types of transitions
}
\description{
function to calculate transition rates at a certain age
}
\examples{
example
}
