% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health5_Simulator.R
\name{health5_simulate_individual_path}
\alias{health5_simulate_individual_path}
\title{function to get the path of 10000 individuals}
\usage{
health5_simulate_individual_path(
  model,
  init_age,
  init_state,
  params,
  gender,
  wave_index,
  cohort = 10000
)
}
\arguments{
\item{model}{1 for no-frailty model, 2 for no-frailty model with a trend, 3 for frailty model}

\item{init_age}{the initial age of the path}

\item{init_state}{the initial state of all individuals}

\item{params}{matrix of estimated parameters to construct the five state model. The rows are beta, gamma_age,
gamma_f, phi (if trend or frailty model), alpha (if frailty model). The columns are 1-12 transition types.}

\item{gender}{gender 1 if female, 0 if male}

\item{cohort}{the number of simulations, set to be 10000}

\item{i}{the wave index}
}
\value{
10000 rows of individual paths in the states, 0 for H, 1 for M, 2 for D, 3 for MD, -1 for Dead
for each row it starts from the initial age as an input, and end at age 110
}
\description{
function to get the path of 10000 individuals
}
\examples{
simulated_individual_path=health5_simulate_individual_path(model='F', init_age=65, init_state=0, params=params_5_frailty, gender=0, wave_index=8, cohort = 10000)
}
