% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health_Data.R
\docType{data}
\name{health5_simulated_path_example}
\alias{health5_simulated_path_example}
\title{10000 simulated paths for males aged 65 in 2012 initially healthy under the 5-state trend model}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 10000 rows and 47 columns.
}
\usage{
health5_simulated_path_example
}
\description{
A matrix containing 10000 simulated paths for males aged 65 in 2012 initially healthy under the 5-state trend model with parameters estimated with US HRS data
Each row of the matrix represent the path of one individual from age 65 to 110
Used as an inbuilt variable to illustrate the statistics functions for the 5-state health model
}
\keyword{datasets}
