% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health5_Statistics.R
\name{health5_stats}
\alias{health5_stats}
\title{Survival Statistics}
\usage{
health5_stats(init_age, init_state, trans_probs)
}
\arguments{
\item{init_age}{integer between 65 and 110 denoting initial age of individual}

\item{init_state}{0 for H state, 1 for M state, 2 for D state, 3 for MD state}

\item{trans_probs}{list of transition probability matrices, generated from \code{health5_get_trans_probs}.}

\item{simulated_path}{matrix containing lifetime path simulations from \code{health5_simulate_paths} function.}
}
\value{
dataframe output containing mean and standard deviation of different statistics
}
\description{
Produces statistics for 5-state model.
}
\examples{
example
}
