% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_Utility.R
\name{mortality_arr_apply}
\alias{mortality_arr_apply}
\title{Apply Functions Over Each Matrix of 3D Array}
\usage{
mortality_arr_apply(X, FUN)
}
\arguments{
\item{X}{3D array}

\item{FUN}{the function to be applied}
}
\value{
3D array of matrices obtained by applying a function
}
\description{
Returns a 3D array of matrices obtained by applying a function over each matrix
of a 3D array.
}
\keyword{internal}
