% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_Data.R
\docType{data}
\name{mortality_AUS_data}
\alias{mortality_AUS_data}
\title{Australian mortality data}
\format{
An object of class \code{demogdata} with the following components:
\describe{
\item{type}{type of object (mortality)}
\item{age}{vector of ages}
\item{year}{vector of years}
\item{rate}{a list containing one or more rate matrices with age rows and year columns}
\item{label}{label of the data.}
\item{pop}{a list of the same form as \code{rate} but containing population numbers
instead of demographic rates}
}
}
\source{
Human Mortality Database \url{http://www.mortality.org/}.
}
\usage{
mortality_AUS_data
}
\description{
Mortality rates of Australians from the Human Mortality Database.
}
\details{
\code{mortality_AUS_data} contains mortality rates of Australians
for the period 1970-2018 and for ages 0-110.
}
\references{
Human Mortality Database (2022). University of California,
Berkeley (USA), and Max Planck Institute for Demographic Research (Germany).
Available at \url{www.mortality.org}.
}
\keyword{datasets}
