% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_CompleteAge.R
\name{mortality_coale_kisker}
\alias{mortality_coale_kisker}
\title{Coale and Kisker Method}
\usage{
mortality_coale_kisker(
  rates,
  ages,
  old_ages,
  type = "central",
  closure_age = 130,
  m_end = 1,
  years = NULL
)
}
\arguments{
\item{rates}{vector, matrix or 3D array of mortality rates with age (on the rows) and
calendar year or cohort (on the columns) and simulation number (3rd dimension)}

\item{ages}{vector of ages for \code{rates}}

\item{old_ages}{vector of old ages for which \code{rates} is to be completed for}

\item{type}{character string representing the type of mortality rate for input and output.
Takes the following values: "central" for central death rates,
"prob" for 1-year death probabilities, "force" for force of mortality}

\item{closure_age}{maximum life span}

\item{m_end}{constant or vector specifying the central death rates at the final age for
each calendar year. If supplied as a vector, vector length and number of
columns for \code{rates} must be equal}

\item{years}{optional vector of years for \code{rates}. If not supplied, then the column names
of \code{rates} will be preserved}
}
\value{
completed mortality rates for all ages and old ages in the same format as
\code{rates}
}
\description{
Implements the Coale and Kisker method of mortality rate completion for old
ages.
}
\examples{
# consider the male mortality rates from the data file 'mortality_AUS_data'
AUS_male_rates <- mortality_AUS_data$rate$male
ages <- mortality_AUS_data$age # 0:110
old_ages <- 91:130

completed_rates <- mortality_coale_kisker(AUS_male_rates, ages, old_ages, type = "central")
}
