% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_Utility.R
\name{mortality_cohort2period}
\alias{mortality_cohort2period}
\title{Convert from Cohort to Period Rates}
\usage{
mortality_cohort2period(cohort_rates)
}
\arguments{
\item{cohort_rates}{matrix or 3D array of mortality rates with age (on the rows) and cohort
(on the columns) and simulation number (3rd dimension)}
}
\value{
associated period mortality rates as a 3D array if \code{cohort_rates} is an array,
or as a matrix otherwise
}
\description{
Converts from cohort to period mortality rates
}
\examples{
# consider the Kannisto completion method on male mortality rates
# from the data file 'mortality_AUS_data'
AUS_male_rates <- mortality_AUS_data$rate$male
ages <- mortality_AUS_data$age # 0:110
old_ages <- 91:130
fitted_ages <- 76:90
completed_rates <- mortality_complete_old_age(
 AUS_male_rates, ages, old_ages,
 method = "kannisto", type = "central",
 fitted_ages = fitted_ages)
# suppose these are rates for cohort starting at age 60
cohort_rates_60 <- completed_rates[as.character(60:130), ]
period_rates <- mortality_cohort2period(cohort_rates_60)
}
