% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_CompleteAge.R
\name{mortality_complete_old_age}
\alias{mortality_complete_old_age}
\title{Mortality Rate Completion}
\usage{
mortality_complete_old_age(
  rates,
  ages,
  old_ages,
  method = "kannisto",
  type = "prob",
  closure_age = 130,
  years = NULL,
  ...
)
}
\arguments{
\item{rates}{vector, matrix or 3D array of mortality rates with age (on the rows) and
calendar year or cohort (on the columns) and simulation number (3rd dimension)}

\item{ages}{vector of ages for \code{rates}}

\item{old_ages}{vector of old ages for which \code{rates} is to be completed for}

\item{method}{character string representing the completion method to be used. Takes the
following values: "CK" for Coale-Kisker, "DG" for Denuit and Goderniaux,
"kannisto" for Kannisto}

\item{type}{character string representing the type of mortality rate for input and output.
Takes the following values: "central" for central death rates,
"prob" for 1-year death probabilities, "force" for force of mortality}

\item{closure_age}{maximum life span}

\item{years}{optional vector of years for \code{rates}. If not supplied, then the column names
of \code{rates} will be preserved}

\item{...}{additional arguments for the chosen completion method. See
\code{\link{mortality_coale_kisker}},
\code{\link{mortality_denuit_goderniaux}},
\code{\link{mortality_kannisto}}}
}
\value{
completed mortality rates for all ages and old ages in the same format as
\code{rates}
}
\description{
Completes mortality rates at old ages.
}
\examples{
# consider the male mortality rates from the data file 'mortality_AUS_data'
AUS_male_rates <- mortality_AUS_data$rate$male
ages <- mortality_AUS_data$age # 0:110
old_ages <- 91:130
# first convert mortality rates to death probabilties
AUS_male_qx <- mortality_rate2rate(AUS_male_rates, from = "central", to = "prob")
# completing mortality rates for old ages
DG_q <- mortality_complete_old_age(AUS_male_qx, ages, old_ages, method = "DG", type = "prob")
CK_q <- mortality_complete_old_age(AUS_male_qx, ages, old_ages, method = "CK", type = "prob")
kannisto_q <- mortality_complete_old_age(AUS_male_qx, ages, old_ages, method = "kannisto", type = "prob", fitted_ages = 80:90)
}
