% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_CompleteAge.R
\name{mortality_kannisto}
\alias{mortality_kannisto}
\title{Kannisto Method of Age Completion}
\usage{
mortality_kannisto(
  rates,
  ages,
  old_ages,
  fitted_ages,
  type = "force",
  closure_age = 130,
  years = NULL
)
}
\arguments{
\item{rates}{vector, matrix or 3D array of mortality rates with age (on the rows) and
calendar year or cohort (on the columns) and simulation number (3rd dimension)}

\item{ages}{vector of ages for \code{rates}}

\item{old_ages}{vector of old ages for which \code{rates} is to be completed for}

\item{fitted_ages}{vector of ages for which model is fitted on}

\item{type}{character string representing the type of mortality rate for input and output.
Takes the following values: "central" for central death rates,
"prob" for 1-year death probabilities, "force" for force of mortality}

\item{closure_age}{maximum life span}

\item{years}{optional vector of years for \code{rates}. If not supplied, then the column names
of \code{rates} will be preserved}
}
\value{
completed mortality rates for all ages and old ages in the same format as
\code{rates}
}
\description{
Implements the Kannisto method of age completion for old ages.
}
\examples{
# consider the male mortality rates from the data file 'mortality_AUS_data'
AUS_male_rates <- mortality_AUS_data$rate$male
ages <- mortality_AUS_data$age # 0:110
old_ages <- 91:130
# fit model on tail end of ages where mortality is still accurate
fitted_ages <- 76:90
completed_rates <- mortality_kannisto(
 AUS_male_rates, ages, old_ages, fitted_ages, type = "central")
}
