% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_Statistics.R
\name{mortality_plot_exp_cfl}
\alias{mortality_plot_exp_cfl}
\title{Plot Curtate Future Lifetime Simulations}
\usage{
mortality_plot_exp_cfl(exp_cfl_rates, years, level = 95)
}
\arguments{
\item{exp_cfl_rates}{matrix of simulated expected curtate future lifetime with simulation number
(on the rows) and calendar year or cohort (on the columns).
Should be generated from \code{\link{exp_cfl}}}

\item{years}{vector of years for \code{exp_cfl_rates}}

\item{level}{desired confidence level with 95\% as default}
}
\description{
Plots simulated expected curtate future lifetime.
}
