% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_PriceOfRisk.R
\name{mortality_rate2survival}
\alias{mortality_rate2survival}
\title{Convert death probabilities to survival function}
\usage{
mortality_rate2survival(
  rates,
  ages,
  from = "prob",
  init_age = NULL,
  years = NULL
)
}
\arguments{
\item{rates}{vector, matrix or 3D array of mortality rates with age
(on the rows) and calendar year or cohort (on the columns) and simulation
number (3rd dimension)}

\item{ages}{vector of ages for \code{rates}}

\item{from}{character string representing the type of mortality rate to be converted
from. Takes the following values: "central" for central death rates, "prob"
for 1-year death probabilities, "force" for force of mortality.}

\item{init_age}{initial age for which the survival function is to be calculated at. If not provided,
the survival function will be calculated for the smallest age supplied in \code{ages}}

\item{years}{optional vector of years for \code{rates}. If not supplied, then the column names
of \code{rates} will be preserved}
}
\value{
associated survival function as a 3D array if \code{rates} is an array, or as a matrix otherwise
}
\description{
Converts mortality rates to the associated survival function.
}
\details{
The survival function has survival time (starting from 0) on the rows.
}
\examples{
# consider the Kannisto completion method on male mortality rates
# from the data file 'mortality_AUS_data'
AUS_male_rates <- mortality_AUS_data$rate$male
ages <- mortality_AUS_data$age # 0:110
old_ages <- 91:130
fitted_ages <- 76:90
completed_rates <- mortality_complete_old_age(
 AUS_male_rates, ages, old_ages, method = "kannisto",
 type = "central", fitted_ages = fitted_ages)
# compute survival function of an individual aged 55
all_ages <- 0:130
surv_func <- mortality_rate2survival(
 completed_rates, ages = all_ages, from = 'central', init_age = 55)
}
