% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_Simulator.R
\name{mortality_sim_cohort_path_realised}
\alias{mortality_sim_cohort_path_realised}
\title{Simulate Cohort Life Path}
\usage{
mortality_sim_cohort_path_realised(
  init_age,
  sex = "F",
  death_probs = NULL,
  closure_age = 130,
  cohort = 1000,
  n_sim = 10000,
  seed = NULL
)
}
\arguments{
\item{init_age}{the initial age of individuals}

\item{sex}{character string denoting the gender of individuals, "F" for female and "M" for male}

\item{death_probs}{a vector of 1-year death probabilities. If not supplied, an M7 age-period-cohort
model will be fitted on \code{mortality_AUS_data} to produce forecasted
death probabilities for a cohort starting at init_age in 2022}

\item{closure_age}{maximum life span}

\item{cohort}{initial cohort size}

\item{n_sim}{number of path simulations}

\item{seed}{integer to be used as the seed for simulation}
}
\value{
a matrix where each row represents the number of individuals still alive
from a given cohort at each age
}
\description{
Simulate Cohort Life Path
}
\examples{
# Simulate life paths for cohort of 1000 females starting at age 60
mortality_sim_cohort_path_realised(init_age = 60, sex = "F")
# Suppose we want to use period 1-yr death probabilities instead
AUS_male_rates <- mortality_AUS_data$rate$male
ages <- mortality_AUS_data$age # 0:110
old_ages <- 91:130
AUS_male_qx <- mortality_rate2rate(AUS_male_rates, from = "central", to = "prob")
kannisto_q <- mortality_complete_old_age(
 AUS_male_qx, ages, old_ages, method = "kannisto",
 type = "prob", fitted_ages = 80:90)
# Consider 100 males aged 55 in the year 2018
qx_55_2018 <- kannisto_q[as.character(55:130), "2018"]
mortality_sim_cohort_path_realised(
 init_age = 55, sex = "M", death_probs = qx_55_2018, cohort = 100)
}
