% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_PriceOfRisk.R
\name{mortality_survival2rate}
\alias{mortality_survival2rate}
\title{Convert survival function to death probabilities}
\usage{
mortality_survival2rate(surv, ages, to = "prob", years = NULL)
}
\arguments{
\item{surv}{vector, matrix or 3D array of the survival function with survival time starting from 0
(on the rows) and calendar year or cohort (on the columns) and simulation number (3rd dimension)}

\item{ages}{vector of desired ages for the resulting 1-year death probabilities}

\item{to}{character string representing the type of mortality rate to be converted
to. Takes the following values: "central" for central death rates, "prob"
for 1-year death probabilities, "force" for force of mortality.}

\item{years}{optional vector of years for \code{surv}. If not supplied, then the column names
of \code{surv} will be preserved}
}
\value{
associated mortality rates in the same format as \code{surv}
}
\description{
Converts the survival function to the associated mortality rates.
}
\examples{
# create survival function for an individual aged 55
AUS_male_rates <- mortality_AUS_data$rate$male
ages <- mortality_AUS_data$age # 0:110
old_ages <- 91:130
fitted_ages <- 76:90
completed_rates <- mortality_complete_old_age(
 AUS_male_rates, ages, old_ages, method = "kannisto",
 type = "central", fitted_ages = fitted_ages)
all_ages <- 0:130
surv_func <- mortality_rate2survival(
 completed_rates, ages = all_ages,
 from = 'central', init_age = 55)
# convert from P to Q measure survival function
# see the section on risk neutral probability
surv_func_Q <- mortality_survivalP2Q(surv_func, method = "wang", lambda = 1.5)
# convert from survival function to mortality rates
central_rates_Q <- mortality_survival2rate(surv_func_Q, 55:130, to = 'central')
}
