% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health_Data.R
\docType{data}
\name{params_5_trend}
\alias{params_5_trend}
\title{Parameters of 5-state Cox Regression Model estimated from US HRS}
\format{
A data frame with 4 rows and 12 variables:
\describe{
\item{trans1}{parameter estimate for H to M transition}
\item{trans2}{parameter estimate for H to D transition}
\item{trans3}{parameter estimate for H to MD transition}
\item{trans4}{parameter estimate for H to Dead transition}
\item{trans5}{parameter estimate for M to MD transition}
\item{trans6}{parameter estimate for M to Dead transition}
\item{trans7}{parameter estimate for D to H transition}
\item{trans8}{parameter estimate for D to M transition}
\item{trans9}{parameter estimate for D to MD transition}
\item{trans10}{parameter estimate for D to Dead transition}
\item{trans11}{parameter estimate for MD to M transition}
\item{trans12}{parameter estimate for MD to Dead transition}
}
}
\source{
\url{https://www.cepar.edu.au/publications/working-papers/multi-state-model-functional-disability-and-health-status-presence-systematic-trend-and-uncertainty}
}
\usage{
params_5_trend
}
\description{
A dataset containing the parameters of the 5-state trend cox regression
model estimated with the 1998-2014 US Health and Retirement Study data.
The rows are beta, gamma_age, gamma_f, phi.
}
\keyword{datasets}
