% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pricing_Valuation.R
\name{plot_convergence}
\alias{plot_convergence}
\title{Plots Convergence of Cashflows}
\usage{
plot_convergence(prices, seed = 9999)
}
\arguments{
\item{seed}{Seed choice for random sampling}

\item{cashflows}{Matrix of simulated cashflow paths}
}
\value{
Convergence Plot
}
\description{
Plots a convergence value for a provided set of cashflows
}
