% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pricing_Valuation.R
\name{plot_distribution}
\alias{plot_distribution}
\title{Plot Distribution of Cashflows}
\usage{
plot_distribution(prices)
}
\arguments{
\item{cashflows}{Matrix of simulated cashflow paths}

\item{convergence}{If convergence = TRUE, plots convergence of price}
}
\value{
Distribution Plot
}
\description{
Plots a histogram for a provided set of cashflows
}
