% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pricing_Valuation.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Plot Distribution of Cashflows}
\usage{
plot_scatter(prices)
}
\arguments{
\item{prices}{Matrix of simulated cashflow paths}
}
\value{
Distribution Plot
}
\description{
Plots a scatter plot for a provided set of cashflows
}
