% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health_Wrapper.R
\name{prob_plots}
\alias{prob_plots}
\title{Create Survival Probability Plots}
\usage{
prob_plots(init_state, init_age, trans_probs)
}
\arguments{
\item{init_state}{integer value
for 3-state model: integer value of 0 or 1, where 0 for healthy state, 1 for disabled state
for 5-state model: 0 for H state, 1 for M state, 2 for D state, 3 for MD state}

\item{init_age}{integer between 65 and 110 denoting current age. This has to the be same as the initial
age used in the generation of transition probability matrices.}

\item{trans_probs}{a list of transition probability matrices, preferably generated from \code{\link[tshm]{get_trans_probs}}.}
}
\value{
line plot denoting Probability of surviving to each different state
}
\description{
Create a plot of survival probabilities to different states.
}
\examples{
trans_probs=get_trans_probs(n_states=5, model_type='F', 
param_file=US_HRS_5, init_age=65, female=0, year = 2012, wave_index = 8, 
latent = 0)
prob_plots(init_state=0, init_age=65, trans_probs=trans_probs)
}
