% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_Simulator.R
\name{sim_indiv_path}
\alias{sim_indiv_path}
\title{Simulate Individual Life Path}
\usage{
sim_indiv_path(
  init_age,
  sex = "F",
  death_probs = NULL,
  closure_age = 130,
  n_sim = 10000,
  seed = NULL
)
}
\arguments{
\item{init_age}{the initial age of individuals}

\item{sex}{character string denoting the gender of individuals, "F" for female and "M" for male}

\item{death_probs}{a vector of 1-year death probabilities. If not supplied, an M7 age-period-cohort
model will be fitted on \code{mortality_AUS_data} to produce forecasted
death probabilities for a cohort starting at init_age in 2022}

\item{closure_age}{maximum life span}

\item{n_sim}{number of path simulations}

\item{seed}{integer to be used as the seed for simulation}
}
\value{
a matrix where each row represents an individual's dead (-1) or alive (0) status
at each age
}
\description{
Simulates the life path of an individual using 1-year death probabilities
}
