% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pricing_Simulator.R
\name{simulate_cf}
\alias{simulate_cf}
\title{Policy Cashflow Simulator}
\usage{
simulate_cf(policy, age = 17, sex = "F", seed = 0, n = 10000)
}
\arguments{
\item{policy}{Policy type to simulate:
`policy object created using a create_policy function}

\item{age}{Initial age of policyholder in years}

\item{sex}{sex of policyholder, \code{sex = "F"} (female), \code{"M"} (male)}

\item{seed}{Seed for random generator}

\item{n}{Number of paths to simulate (Monte-Carlo method)}
}
\value{
Matrix of cash flow vectors for each simulated path
}
\description{
Simulate cash flows using Monte-Carlo methods for various policies
}
\examples{
ap <- create_policy_AP(400000, 60000)
cf <- cashflow(policy = ap, age = 65, sex = "M", n = 1000)
}
