% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health_Transition.R
\name{simulate_health_state_paths}
\alias{simulate_health_state_paths}
\title{Simulate cohort life path}
\usage{
simulate_health_state_paths(
  trans_probs,
  init_age,
  init_state = 0,
  cohort = 1e+05
)
}
\arguments{
\item{trans_probs}{a list of transition probability matrices, preferably generated from \code{\link[tshm]{get_trans_probs}}.}

\item{init_age}{integer between 65 and 110 denoting current age. This has to the be same as the initial
age used in the generation of transition probability matrices.}

\item{init_state}{0 for healthy, 1 for disabled}

\item{cohort}{integer (default 10000) denoting number of people in the simulation}
}
\value{
a matrix where each row represents a new individual, and the columns represent
the individual's movement through each state.

-1 (death) is absorbing, so if an individual enters that state, the rest of the row will be -1.
}
\description{
Simulates the path each life takes in an initial cohort using transition probabilities
}
\examples{
example
}
