% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health_Transition.R
\name{simulate_life_table}
\alias{simulate_life_table}
\title{Simulate life table for frailty model}
\usage{
simulate_life_table(
  n_states,
  param_file,
  age,
  gender,
  year = 2012,
  wave_index = 8,
  n_sim = 10000,
  mean = TRUE
)
}
\arguments{
\item{param_file}{string for file path of parameter file OR a tibble/dataframe of parameters}

\item{age}{integer denoting age of policy holder}

\item{gender}{takes values 1 or 0, where 1 indicates policyholder is female}

\item{year}{integer denoting current year}

\item{wave_index}{the wave index}

\item{n_sim}{number of simulations}

\item{mean}{return expected life table}

\item{model_type}{string that selects model type; S for Static, T for Trend and F for Frailty}

\item{latent}{initial value of latent factor, normally take the value 0}
}
\value{
list of life tables (default) or expected life table (mean == TRUE)
}
\description{
Simulate life table for frailty model starting from initial age to
max age of 110 for 3 state model
}
\examples{
example

}
