% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_Simulator.R
\name{simulate_path_mortality}
\alias{simulate_path_mortality}
\title{Simulate Individual Life Path}
\usage{
simulate_path_mortality(
  init_age,
  sex = "F",
  death_probs = NULL,
  closure_age = 130,
  n_sim = 10000
)
}
\arguments{
\item{init_age}{the initial age of the path}

\item{sex}{character string denoting the gender of individuals, "F" for female and "M" for male}

\item{death_probs}{a vector of 1-year death probabilities. If not supplied, an M7 age-period-cohort
model will be fitted on \code{mortality_AUS_data} to produce forecasted
death probabilities for a cohort starting at init_age in 2022}

\item{closure_age}{maximum life span}

\item{n_sim}{integer (default 10000) denoting number of path simulations, which is equivalent
to the number of people in the simulation}
}
\value{
a matrix where each row represents an individual's dead (-1) or alive (0) status
across the years
}
\description{
Simulates the life path of an individual using 1-year death probabilities
}
\examples{

}
