% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health5_Statistics.R
\name{stats_produce_5}
\alias{stats_produce_5}
\title{function to produce the mean and variance of a list of values}
\usage{
stats_produce_5(input)
}
\arguments{
\item{input}{the list of values to be studied}
}
\value{
mean and variance of the
}
\description{
function to produce the mean and variance of a list of values
}
\examples{
stats_first_time_leave_H=stats_produce_5(first_time_stats(simulated_path,0))
}
