% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_PriceOfRisk.R
\name{survivalP2Q}
\alias{survivalP2Q}
\title{Survival Function Transformation}
\usage{
survivalP2Q(StP, method, lambda)
}
\arguments{
\item{StP}{vector, matrix or 3D array of the survival function under the P-measure with
survival time (on the rows) and calendar year or cohort (on the columns) and simulation
number (3rd dimension)}

\item{method}{character string representing the distortion risk measure to be used. See "Details".}

\item{lambda}{parameter associated with the distortion risk measure}
}
\value{
the transformed survival function under the specified Q-measure
}
\description{
Transforms the survival function from the real world P-measure to the risk-neutral
Q-measure according to the specified risk-neutral principles.
}
\details{
The risk-neutral principles and their corresponding strings and valid lambda range
are as follows:
\itemize{
\item "wang", \eqn{\lambda \ge 0}: Wang Transform,
\item "ph", \eqn{\lambda \ge 1}: Proportional Hazard Transform
\item "dp", \eqn{\lambda \ge 1}: Dual-power Transform
\item "gp", \eqn{0 \le \lambda \le 1}: Gini Principle
\item "dadp", \eqn{0 \le \lambda \le 1}: Denneberg's Absolute Deviation Principle
\item "exp", \eqn{\lambda > 0}: Exponential Transform
\item "log", \eqn{\lambda > 0}: Logarithmic Transform
\item "canon", \eqn{\lambda > 0}: Univariate Canonical Valuation
\item "esscher", \eqn{\lambda > 0}: Esscher Transform
}

The first seven principles are distortion risk measures and act on the survival function.
The last two principles act on the probability density function.
}
