% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health5_Statistics.R
\name{total_time_stats_5}
\alias{total_time_stats_5}
\title{function to get the total time in different states for a number of individuals}
\usage{
total_time_stats_5(simulated_path, state)
}
\arguments{
\item{simulated_path}{the simulated path of individuals from the function simulate_individual_path}

\item{state}{0 for total time in H state
1 for total time in M state
2 for total time in D state
3 for total time in MD state
-1 for total time in dead state
4 for total time alive or not in dead state}
}
\value{
a column that consists the total time in different states for a number of individuals
}
\description{
function to get the total time in different states for a number of individuals
}
\examples{
total_time_alive=total_time_stats_5(simulated_individual_path, 4)
}
