% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health5_Transition.R
\name{transition_rate_5}
\alias{transition_rate_5}
\title{function to calculate transition rates at a certain age}
\usage{
transition_rate_5(params, age, gender, i, latent, model)
}
\arguments{
\item{params}{matrix of estimated parameters to construct the five state model. The rows are beta, gamma_age,
gamma_f, phi (if trend or frailty model), alpha (if frailty model). The columns are 1-12 transition types.}

\item{age}{age of the individual}

\item{gender}{gender 1 if female, 0 if male}

\item{i}{the wave index}

\item{latent}{initial value of latent factor, normally take the value 0}

\item{model}{1 for no-frailty model, 2 for no-frailty model with a trend, 3 for frailty model}
}
\value{
12 times 1 vector of transition rates for the 12 types of transitions
}
\description{
function to calculate transition rates at a certain age
}
\examples{
transition_rates=transition_rate_5(params=params_5_frailty, age=65, gender=0, i=8, latent=0, model=3)
}
