% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pricing_Valuation.R
\name{value_policy}
\alias{value_policy}
\title{Valuation of Policy}
\usage{
value_policy(policy, cashflows, seed = 0)
}
\arguments{
\item{policy}{Policy object containing necessary parameters (see create_policy_)}

\item{cashflows}{(Optional) Matrix of cashflows (see simulate_cf), if 'cashflows' is not
provided, function will simulate cashflows for policy}

\item{seed}{Seed choice for random sampling}
}
\value{
'pol_value' object: contains summary summary statistics and
convergence + distribution plots
}
\description{
Conducts valuation of a given policy, providing summary statistics and
convergence / distribution plots
}
\examples{
la <- create_policy_LA(60000, 5, 0, 0.01)
cf_la <- simulate_cf(la)
val_la <- value_policy(la, cf_la)
}
