% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health3_Statistics.R
\name{health3_dfl}
\alias{health3_dfl}
\title{Disabled Future Lifetime}
\usage{
health3_dfl(init_age, init_state, trans_probs = NULL, simulated_path = NULL)
}
\arguments{
\item{init_age}{integer between 65 and 110 denoting age of individual}

\item{init_state}{0 for healthy, 1 for disabled}

\item{trans_probs}{a list of transition probability matrices, preferably generated from
\code{health3_get_trans_probs}.}

\item{simulated_path}{a matrix of lifetime paths simulations, generated by \code{health3_simulate_paths}.}
}
\value{
numeric output for average future lifetime in disabled state and its standard
deviation.
}
\description{
Calculates the average future lifetime spent in disabled state and the standard
deviation by simulating life time paths given a list of transition probability
matrices. Function uses either list of transition probability matrices, or a
simulated path. If both are provided, then the simulated path will be used.
NOTE: THIS FUNCTION SHOULD ONLY BE USED WITH STATIC AND TREND MODELS. USE
\code{health3_dflF} FOR FRAILTY MODEL.
}
\examples{
example
}
