% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health3_Statistics.R
\name{health3_hfl}
\alias{health3_hfl}
\title{Healthy Future lifetime}
\usage{
health3_hfl(init_age, init_state, trans_probs = NULL, simulated_path = NULL)
}
\arguments{
\item{init_age}{integer between 65 and 110 denoting initial age of individual}

\item{init_state}{0 for healthy, 1 for disabled}

\item{trans_probs}{list of transition probability matrices; ideally generated from
\code{health3_get_trans_probs}.}

\item{simulated_path}{a matrix of lifetime paths simulations, generated by \code{health3_simulate_paths}.}
}
\value{
Numeric output of expected time and standard deviation of time spent in healthy state.
}
\description{
Calculates the expected future lifetime spent and its standard deviation
in the healthy state. Function uses either transition probability matrices, or
a simulated lifetime paths. If both are provided, it will use the simulated lifetime
provided.
NOTE: USE \code{health3_hflF} FOR FRAILTY MODEL.
}
\examples{
example
}
