% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health3_LifeTable.R
\name{health3_simulate_life_table}
\alias{health3_simulate_life_table}
\title{Create life table (frailty model)}
\usage{
health3_simulate_life_table(
  init_age,
  female,
  year,
  param_file,
  init_state,
  n_sim,
  cohort,
  mean
)
}
\arguments{
\item{init_age}{integer between 65 and 110 denoting initial age of individual}

\item{female}{1 for female, 0 for male}

\item{year}{integer denoting year of individual}

\item{param_file}{file containing parameters of cox regression model}

\item{init_state}{0 for healthy, 1 for disabled}

\item{n_sim}{number of latent factor paths to simulate}

\item{cohort}{cohort size of lifetable}

\item{mean}{FALSE to return list of lifetables, TRUE to return expected lifetable}
}
\value{
Dataframe containing life table
}
\description{
Creates a large number of life tables simulated from different latent factor paths
and averages them to create 1 life table output.
}
\examples{
example
}
