% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health3_Statistics.R
\name{health3_survival_stats}
\alias{health3_survival_stats}
\title{Survival Statistics}
\usage{
health3_survival_stats(
  init_age,
  init_state,
  trans_probs = NULL,
  simulated_path = NULL
)
}
\arguments{
\item{init_age}{integer between 65 and 110 denoting initial age of individual}

\item{init_state}{0 for healthy, 1 for disabled}

\item{trans_probs}{list of transition probability matrices, generated from \code{health3_get_trans_probs}.}

\item{simulated_path}{matrix containing lifetime path simulations from \code{health3_simulate_paths} function.}
}
\value{
dataframe output containing mean and standard deviation of different statistics
}
\description{
Produces statistics including: total expected lifetime, healthy lifetime,
disabled lifetime, onset of disability (if initial state is healthy). It uses
either transition probability matrices or a simulated paths.
NOTE: USE \code{health3_survival_statsF} FOR FRAILTY MODEL.
}
\examples{
example
}
