% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health3_Statistics.R
\name{health3_time_to_disabled}
\alias{health3_time_to_disabled}
\title{Time until onset of disability (conditional on being disabled)}
\usage{
health3_time_to_disabled(init_age, trans_probs = NULL, simulated_path = NULL)
}
\arguments{
\item{init_age}{integer between 65 and 110 denoting initial age of individual}

\item{trans_probs}{a list of transition probability matrices, preferably generated from
\code{health3_get_trans_probs}.}

\item{simulated_path}{a matrix of lifetime paths simulations, generated by \code{health3_simulate_paths}.}
}
\value{
numeric denoting the average time until individual becomes disabled and its
standard deviation.
}
\description{
Uses simulation to produce an average time until a healthy individual becomes disabled
during their life time. Note that this is conditional on the individual becoming disabled,
so it doesn't factor in life simulations where the individual is always healthy. Function
requires either list of transition probability matrices, or simulated lifetime paths.
If both are given, then the simulated lifetime paths is used.
NOTE: USE \code{health3_time_to_disabledF} for frailty model.
}
\examples{
example
}
