% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health5_Transition.R
\name{health5_get_trans_probs}
\alias{health5_get_trans_probs}
\title{the function to get a full list of transition probability matrices from the
initial age to age 110}
\usage{
health5_get_trans_probs(
  model_type,
  param_file,
  init_age,
  female,
  wave_index,
  latent
)
}
\arguments{
\item{model_type}{S for static model, T for trend model, F for frailty model}

\item{param_file}{matrix of estimated parameters to construct the five state model. Generally,
use US_HRS_5 for 5 state model.}

\item{init_age}{the initial age of the transition probability matrices}

\item{female}{female 1 if female, 0 if male}

\item{wave_index}{the wave index}

\item{latent}{initial value of latent factor, normally take the value 0}
}
\value{
a list of 5 times 5 transition probability matrices, from the
initial age to age 110
}
\description{
the function to get a full list of transition probability matrices from the
initial age to age 110
}
\examples{
example
}
