% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health5_Simulator.R
\name{health5_simulate_life_table}
\alias{health5_simulate_life_table}
\title{the function to get n_sim number of simulated lifetables for Frailty model}
\usage{
health5_simulate_life_table(
  model_type,
  param_file,
  female,
  wave_index,
  latent,
  init_age,
  init_state,
  n_sim,
  cohort,
  mean
)
}
\arguments{
\item{model_type}{choose F for Frailty model}

\item{param_file}{matrix of estimated parameters to construct the five state model. Generally,
use US_HRS_5 for 5 state model.}

\item{female}{female 1 if female, 0 if male}

\item{wave_index}{the wave index = (interview year - 1998)/2 + 1}

\item{latent}{initial value of latent factor, normally take the value 0}

\item{init_age}{the initial age of the life table}

\item{init_state}{0 for H state, 1 for M state, 2 for D state, 3 for MD state}

\item{n_sim}{the number of simulations}

\item{cohort}{number of people at the beginning of the life table}

\item{mean}{TRUE to return expected life table, FALSE to return all simulated life tables}
}
\value{
a list of n_sim number of life table matrices when mean=FALSE
for each matrix, the row represents the age from the input initial age to 110, and the columns are states H M D MD Dead
or the mean life table when mean=TRUE
}
\description{
the function to get n_sim number of simulated lifetables for Frailty model
}
\examples{
example
}
