% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health5_Simulator.R
\name{health5_simulate_paths}
\alias{health5_simulate_paths}
\title{function to get the path of set number of individuals}
\usage{
health5_simulate_paths(list_trans_probs, init_age, init_state, cohort)
}
\arguments{
\item{list_trans_probs}{a list of transition probability matrices, preferably generated from \code{get_trans_probs}.}

\item{init_age}{the initial age of the path}

\item{init_state}{the initial state of all individuals}

\item{cohort}{the number of simulated cohorts}
}
\value{
rows of individual paths in the states, 0 for H, 1 for M, 2 for D, 3 for MD, -1 for Dead
for each row it starts from the initial age as an input, and end at age 110
}
\description{
function to get the path of set number of individuals
}
\examples{
example
}
