% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health5_Statistics.R
\name{health5_stats}
\alias{health5_stats}
\title{Survival Statistics}
\usage{
health5_stats(
  model_type,
  init_age,
  init_state,
  trans_probs = NULL,
  simulated_path = NULL,
  female = NULL,
  year = NULL,
  wave_index = NULL,
  latent = NULL,
  param_file = NULL,
  n = 1000
)
}
\arguments{
\item{model_type}{S for static model, T for trend model, F for frailty model}

\item{init_age}{integer between 65 and 110 denoting initial age of individual. This needs to be same
initial age used in generation of \code{trans_probs} or \code{simulated_path}}

\item{init_state}{0 for healthy, 1 for disabled}

\item{trans_probs}{a list of transition probability matrices, preferably generated from
\code{get_trans_probs}, only needed for static and trend models.}

\item{simulated_path}{the simulated path of individuals from the function \code{simulate_health_state_paths}}

\item{female}{0 for male, 1 for female, compulsory variable for frailty model}

\item{year}{integer indicating current year, compulsory variable for frailty model}

\item{wave_index}{the wave index = (interview year - 1998)/2 + 1, compulsory variable for frailty model}

\item{latent}{initial value of latent factor, normally take the value 0, compulsory variable for frailty model}

\item{param_file}{string for file path of parameter file OR a tibble/dataframe of parameters, compulsory variable for frailty model}

\item{n}{integer denoting number of unique latent factor simulations}

\item{state}{0 for first time leaving H state, only useful when initial state is 0
1 for first time entering M state
2 for first time entering D state
3 for first time entering MD state
-1 for first time entering the dead state}
}
\value{
dataframe output containing mean and standard deviation of different statistics
}
\description{
Produces statistics for 5-state model.
}
\examples{
example
}
