% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health5_Statistics.R
\name{health5_stats_produce}
\alias{health5_stats_produce}
\title{function to produce the mean and variance of a list of values}
\usage{
health5_stats_produce(input)
}
\arguments{
\item{input}{the list of values to be studied}
}
\value{
mean and variance of the list of values
}
\description{
function to produce the mean and variance of a list of values
}
\examples{
stats=health5_stats_produce(health5_first_time_stats(health5_simulated_path_example, 0))
}
