% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health5_Statistics.R
\name{health5_total_time_stats}
\alias{health5_total_time_stats}
\title{function to get the total time in different states for a number of individuals}
\usage{
health5_total_time_stats(simulated_path, state)
}
\arguments{
\item{simulated_path}{the simulated path of individuals from the function \code{simulate_health_state_paths}}

\item{state}{0 for total time in H state
1 for total time in M state
2 for total time in D state
3 for total time in MD state
-1 for total time in dead state
4 for total time alive or not in dead state}
}
\value{
a column that consists the total time in different states for a number of individuals
}
\description{
function to get the total time in different states for a number of individuals
}
\examples{
total_time_alive=health5_total_time_stats(health5_simulated_path_example, 4)
}
