% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pricing_Valuation.R
\name{plot_distribution}
\alias{plot_distribution}
\title{Plot Distribution of Cashflows}
\usage{
plot_distribution(prices)
}
\arguments{
\item{prices}{Matrix of simulated cashflow paths}
}
\value{
Distribution Plot
}
\description{
Plots a histogram for a provided set of cashflows
}
