% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_Simulator.R
\name{sim_cohort_path_realised}
\alias{sim_cohort_path_realised}
\title{Simulate Cohort Life Path}
\usage{
sim_cohort_path_realised(
  init_age,
  sex = "F",
  death_probs = NULL,
  closure_age = 130,
  cohort = 1000,
  n_sim = 10000,
  seed = NULL
)
}
\arguments{
\item{init_age}{the initial age of individuals}

\item{sex}{character string denoting the gender of individuals, "F" for female and "M" for male}

\item{death_probs}{a vector of 1-year death probabilities. If not supplied, an M7 age-period-cohort
model will be fitted on \code{mortality_AUS_data} to produce forecasted
death probabilities for a cohort starting at init_age in 2022}

\item{closure_age}{maximum life span}

\item{cohort}{initial cohort size}

\item{n_sim}{number of path simulations}

\item{seed}{integer to be used as the seed for simulation}
}
\value{
a matrix where each row represents the number of individuals still alive
from a given cohort at each age
}
\description{
Simulate Cohort Life Path
}
