% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pricing_Simulator.R
\name{simulate_cf}
\alias{simulate_cf}
\title{Policy Cashflow Simulator}
\usage{
simulate_cf(
  policy,
  age = 65,
  sex = "F",
  seed = 0,
  n = 100,
  state = NULL,
  econ_var = NULL
)
}
\arguments{
\item{policy}{Policy type to simulate:
`policy object created using a create_policy function}

\item{age}{Initial age of policyholder in years}

\item{sex}{sex of policyholder, \code{sex = "F"} (female), \code{"M"} (male)}

\item{seed}{Seed for random generator}

\item{n}{Number of paths to simulate (Monte-Carlo method)}

\item{state}{Simulated state matrix via Health-State / Aggregate Mortality}

\item{econ_var}{Simulated economic variables via Economic Scenario Generator}
}
\value{
Matrix of cash flow vectors for each simulated path
}
\description{
Simulate cash flows using Monte-Carlo methods for various policies
}
\examples{
ap <- create_policy_AP(400000, 60000)
cf <- simulate_cf(policy = ap, age = 65, sex = "M", n = 1000)
}
