% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health_Wrapper.R
\name{simulate_health_state_paths}
\alias{simulate_health_state_paths}
\title{Simulate cohort life path}
\usage{
simulate_health_state_paths(
  trans_probs,
  init_age,
  init_state = 0,
  cohort = 10000
)
}
\arguments{
\item{trans_probs}{a list of transition probability matrices, preferably generated from \code{\link[tshm]{get_trans_probs}}.}

\item{init_age}{integer between 65 and 110 denoting current age. This has to the be same as the initial
age used in the generation of transition probability matrices.}

\item{init_state}{integer value
for 3-state model: integer value of 0 or 1, where 0 for healthy state, 1 for disabled state
for 5-state model: 0 for H state, 1 for M state, 2 for D state, 3 for MD state}

\item{cohort}{integer (default 10000) denoting number of people in the simulation}
}
\value{
a matrix where each row represents a new individual, and the columns represent
the individual's movement through each state.
for 3-state model: 0 for healthy state, 1 for disabled state, -1 for death state
for 5-state model: 0 for H state, 1 for M state, 2 for D state, 3 for MD state, -1 for death state
-1 (death) is absorbing, so if an individual enters that state, the rest of the row will be -1.
}
\description{
Simulates the path each life takes in an initial cohort using transition probabilities
}
\examples{
trans_probs=get_trans_probs(n_states=5, model_type='F', 
param_file=US_HRS_5, init_age=65, female=0, year = 2012, wave_index = 8, 
latent = 0)
simulated_path <- simulate_health_state_paths(trans_probs, init_age=65, 
init_state = 0, cohort = 10000)
}
