% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health_Wrapper.R
\name{simulate_life_table}
\alias{simulate_life_table}
\title{Simulate life table for frailty model}
\usage{
simulate_life_table(
  model_type,
  param_file,
  female,
  year = 2012,
  wave_index = 8,
  latent = 0,
  init_age,
  init_state = 0,
  n_sim = 100,
  cohort = 1,
  mean = FALSE
)
}
\arguments{
\item{model_type}{string that selects model type; S for Static, T for Trend and F for Frailty}

\item{param_file}{string for file path of parameter file OR a tibble/dataframe of parameters}

\item{female}{takes values 1 or 0, where 1 indicates policyholder is female}

\item{year}{integer denoting current year}

\item{wave_index}{the wave index}

\item{latent}{initial value of latent factor, normally take the value 0
#' @param init_age
integer denoting current age}

\item{init_state}{integer value of 0 or 1, where 0 indicates healthy and 1 indicates disabled}

\item{n_sim}{number of simulations}

\item{cohort}{number of people at the beginning of the life table}

\item{mean}{TRUE tp return expected life table, FALSE to return all simulated life tables}
}
\value{
list of life tables (default) or expected life table (mean == TRUE)
}
\description{
Simulate life table for frailty model starting from initial age to
max age of 110 for 3 state model
}
\examples{
example

}
