% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health_Wrapper.R
\name{simulate_life_table}
\alias{simulate_life_table}
\title{Simulate life table for frailty model}
\usage{
simulate_life_table(
  n_states,
  model_type,
  param_file,
  init_age,
  female,
  year = 2012,
  init_state = 0,
  wave_index = 8,
  latent = 0,
  n_sim = 100,
  cohort = 1e+05,
  mean = FALSE
)
}
\arguments{
\item{n_states}{take values 3 or 5, use 3 for 3-state model, and 5 for 5-state model}

\item{model_type}{select F for Frailty model}

\item{param_file}{string for file path of parameter file OR a tibble/dataframe of parameters}

\item{init_age}{integer denoting current age}

\item{female}{takes values 1 or 0, where 1 indicates policyholder is female}

\item{year}{integer denoting current year}

\item{init_state}{integer value
for 3-state model: integer value of 0 or 1, where 0 for healthy state, 1 for disabled state
for 5-state model: 0 for H state, 1 for M state, 2 for D state, 3 for MD state}

\item{wave_index}{the wave index = (interview year - 1998)/2 + 1}

\item{latent}{initial value of latent factor, normally take the value 0}

\item{n_sim}{number of simulations}

\item{cohort}{number of people at the beginning of the life table}

\item{mean}{TRUE to return expected life table, FALSE to return all simulated life tables}
}
\value{
list of life tables (default) or expected life table (mean == TRUE)
}
\description{
Simulate life table for for 3 state or 5 state frailty model starting from initial age to
max age of 110
}
\examples{
lifetable_simulated <- simulate_life_table(n_states=5, model_type='F', 
param_file=US_HRS_5, init_age=65, female=0, year = 2012, init_state = 0, 
wave_index = 8,latent=0,n_sim=100,cohort=100,mean=FALSE)

}
