% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_Statistics.R
\name{surv}
\alias{surv}
\alias{psurv}
\alias{qsurv}
\title{The Survival Function}
\usage{
psurv(surv_fun, surv_time)

qsurv(surv_fun, surv_prob)
}
\arguments{
\item{surv_fun}{vector of survival function}

\item{surv_time}{vector of survival times}

\item{surv_prob}{vector of survival probabilities}
}
\description{
Distribution and quantile function  of the survival function.
}
\details{
Linear interpolation is performed between the discrete points of the survival function
}
